refactor: change way to initialize logging for deployment
