+++ deploy/__init__.py
@@ -1,29 +1,15 @@
 """
 Package for deployment
 """
-import logging.config
 import logging
-import os
 
 from ansible import AnsibleExecutor
 from deploy.deploy_const import DEVELOPMENT_MODE, PRODUCTION_MODE
 from deploy.jobs import *
+from deploy.utils import create_directory
 from settings import config
 
 logger = logging.getLogger('ansible_deploy')
-
-
-def configure_deploy_logging_locally(logger_file: str):
-    """
-    Configures deploy logging file locally
-
-    Args:
-        logger_file: path of the logger file (expected that directory to this file is already exist)
-    """
-    deploy_log_config = config.logging_ansible_deploy
-    if not deploy_log_config['handlers']['service_file']['filename']:
-        deploy_log_config['handlers']['service_file']['filename'] = logger_file
-    logging.config.dictConfig(config=deploy_log_config)
 
 
 async def perform_deployment(deploy_mode: str, local_output_dir: str):
@@ -33,10 +19,9 @@ async def perform_deployment(deploy_mode: str, local_output_dir: str):
     Args:
         deploy_mode: mode of deployment
         local_output_dir: path to an existing local directory to be used:
-                              - for deployment log files
                               - as ansible-runner's private data directory
     """
-    configure_deploy_logging_locally(logger_file=os.path.join(local_output_dir, 'ansible-deploy.log'))
+    create_directory(local_output_dir)
 
     assert deploy_mode in (PRODUCTION_MODE, DEVELOPMENT_MODE), \
         f"Only two modes of deployment is allowed: '{DEVELOPMENT_MODE}' and '{PRODUCTION_MODE}'"
