+++ deploy/deploy_const.py
@@ -3,7 +3,7 @@ Deploy constants
 """
 from pathlib import Path
 
-PROJECT_ROOT_PATH = str(Path(__file__).parent.parent.resolve())
+PROJECT_ROOT_PATH = Path(__file__).parent.parent.resolve()
 
 # allowed modes
 DEVELOPMENT_MODE = 'development'
