+++ deploy/run.py
@@ -2,6 +2,8 @@
 Automatic deployment with Ansible and ansible-runner
 """
 import asyncio
+import logging
+import logging.config
 import os
 
 from concurrent.futures import ThreadPoolExecutor
@@ -11,21 +13,31 @@ from deploy.deploy_const import PROJECT_ROOT_PATH
 from settings import config
 
 
-def create_directory(dir_path: str):
-    """ Creates a directory on the path if it is not created """
+def configure_deploy_logging_locally(logger_file: str):
+    """
+    Configures deploy logging file locally
+
+    Args:
+        logger_file: path of the logger file
+    """
+    dir_path = os.path.dirname(logger_file)
     try:
         os.makedirs(dir_path)
         print(f"Directory {dir_path} successfully created")
     except FileExistsError:
         print(f"Directory {dir_path} already exist")
 
+    deploy_log_config = config.logging_ansible_deploy
+    if not deploy_log_config['handlers']['service_file']['filename']:
+        deploy_log_config['handlers']['service_file']['filename'] = logger_file
+    logging.config.dictConfig(config=deploy_log_config)
+
 
 if __name__ == '__main__':
-    tmp_directory = os.path.join(PROJECT_ROOT_PATH, 'tmp/')
-    create_directory(dir_path=tmp_directory)
+    configure_deploy_logging_locally(logger_file=os.path.join(PROJECT_ROOT_PATH.joinpath('tmp/'), 'ansible-deploy.log'))
 
     asyncio.new_event_loop().set_default_executor(ThreadPoolExecutor(max_workers=2))
     asyncio.run(
         perform_deployment(deploy_mode=config.deploy.mode,
-                           local_output_dir=tmp_directory)
+                           local_output_dir=PROJECT_ROOT_PATH.joinpath('tmp/ansible'))
     )
